ref: Doc comments, Move scope to seperate file
